cp8-PF/apirestfull/requirements.txt
aio-pika==9.3.1
aiormq==6.7.7
annotated-types==0.6.0
anyio==3.7.1
asynctest==0.13.0
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.104.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
multidict==6.0.4
pamqp==3.2.1
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
yarl==1.9.3
